src/main/java/com/rhyno/demoapi/project/ProjectController.java
package com.rhyno.demoapi.project;

import com.rhyno.demoapi.project.model.Project;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1/projects")
public class ProjectController {
    private final ProjectService projectService;

    public ProjectController(ProjectService projectService) {
        this.projectService = projectService;
    }

    @GetMapping("/{projectId}")
    public Project getProject(@PathVariable(name = "projectId") String id) {
        return projectService.getProject(id);
    }
}

